Move the SIMD headers from 'config' to 'linalg'.
