+++ linalg/simd.hpp
