+++ linalg/simd/m512.hpp
@@ -14,7 +14,7 @@
 
 #ifdef __AVX512F__
 
-#include "../tconfig.hpp"
+#include "../../config/tconfig.hpp"
 #include <x86intrin.h>
 
 namespace mfem
