+++ linalg/simd/qpx.hpp
