+++ linalg/simd/vsx.hpp
