+++ linalg/simd/x86.hpp
